fix: handle edge case when alternative file exists but the package does not
This happens for openjdk-25, the package is not available for i686, but
the script still tries to parse it and add all the binaries, but it
fails to find the binary. I don't think we should do the validation here
in command-not-found, but rather deal with it in repository consistency
checks in termux-packages.

Also instead of quietly ignoring the alternatives file, it'll split out
a warning, so in future if anything else goes wrong due to this, at
least the logs do contain some hint